+++ UUEPULEFXVI_3TTROTL/UUEPULEFXVI_3TTROTL.vcxproj
@@ -150,6 +150,17 @@
   <ItemGroup>
     <ClCompile Include="main.cpp" />
   </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="Animation.h" />
+    <ClInclude Include="Entity.h" />
+    <ClInclude Include="Bullet.h" />
+    <ClInclude Include="constants.h" />
+    <ClInclude Include="Map.h" />
+    <ClInclude Include="Pickup.h" />
+    <ClInclude Include="Player.h" />
+    <ClInclude Include="Walls.h" />
+    <ClInclude Include="Zombie.h" />
+  </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
   </ImportGroup>
